Fix OpenSSL compilation on macOS; use OpenSSL 3.1
This PR adds macOS support for OpenSSL 3.1 on both Intel and Apple Silicon.

Without this fix, Ruby scripts compiled for macOS that make HTTPS requests result in OpenSSL errors such as:

```
certificate verify failed (EE certificate key too weak) (OpenSSL::SSL::SSLError)
```

and

```
unable to find ecdh parameters
```

To fix this, I updated the OpenSSL configuration to use the same settings that Homebrew uses here:

https://github.com/Homebrew/homebrew-core/blob/HEAD/Formula/o/openssl@3.rb#L61-L68

I kept the `no-shared` option since we want static libraries.

As for the `no-asm` option that I removed, according to the OpenSSL repo, it's meant to be used for debugging/troubleshooting, and not for production use:

https://github.com/openssl/openssl/blob/b07107e31149bf870bc1ae17e59444859fe4e23a/INSTALL.md?plain=1#L591

And `enable-rc5` is disabled by default, and is for RC5 symmetric encryption using the EVP_CIPHER API:

https://github.com/openssl/openssl/blob/b07107e31149bf870bc1ae17e59444859fe4e23a/doc/man7/EVP_CIPHER-RC5.pod#L11

Both `no-asm` and `enable-rc5` were added in the motor-admin fork to add support for Apple Silicon, but I don't think they're needed. Here's the original PR:

https://github.com/motor-admin/ruby-packer/commit/99023b49a569abd07df6670aa2f950ccc1f22354

My paid Ruby CLI compiled with rubyc makes several HTTPS requests and it all works fine with the settings in this PR, on both Intel and Apple Silicon Macs. Hundreds of customers on various macOS versions and Mac models have used it without issues.

I also removed `-Wno-error=implicit-function-declaration` from CFLAGS because that's a hack that's often recommended when Ruby installation fails on macOS, but is not necessary, especially with newer Ruby versions like 3.1.3. In my 13 years of using Ruby, not once have I had to use that flag on macOS.