Fix compilation issue with CLT > 15.1
The error was incompatible pointer to integer conversion passing 'struct stat *' to parameter of type 'int' [-Wint-conversion] because the access method expects an int (the mode) as its second argument, but it was passing the &buf. So then through some searching, I found that you can extract the mode from the stat struct via buf.st_mode.